requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0,<4.0.0
celery>=5.2.0,<6.0.0
redis>=4.3.0,<5.0.0
django-celery-beat>=2.4.0,<3.0.0
drf-yasg>=1.21.0,<2.0.0
requests>=2.28.0,<3.0.0
gunicorn>=20.1.0,<21.0.0
python-dateutil>=2.8.2,<3.0.0
psycopg2-binary>=2.9.5,<3.0.0
dj-database-url>=2.0.0,<3.0.0 